introduce own shortcuts module
